Move song content to another thread
