+++ app/src/main/java/com/example/repertoire/SongActivity.kt
@@ -1,12 +1,10 @@
 package com.example.repertoire
 
-import android.net.Uri
 import android.os.Bundle
+import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.Observer
 import kotlinx.android.synthetic.main.activity_song.*
-import java.io.BufferedReader
-import java.io.IOException
-import java.io.InputStreamReader
 
 class SongActivity : AppCompatActivity() {
     companion object {
@@ -18,35 +16,18 @@ class SongActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_song)
         supportActionBar?.hide()
+        val observer = Observer<String> { content -> song_text_view.text = content }
+        songViewModel.songContent.observe(this, observer)
 
         val bundle = intent.extras!!
-        val songName = bundle[SONG_NAME] as String
-        val songUriAsString = bundle[SONG_URI_AS_STRING] as String
-        val songContent = readSongFile(Uri.parse(songUriAsString)) // TODO: This should not be on the main thread
-        song_text_view.text = "$songName $songContent"
+        song = Song (
+            name = bundle[SONG_NAME] as String,
+            uri = bundle[SONG_URI_AS_STRING] as String
+        )
+        songViewModel.setSongContent(song)
     }
 
 
-    private fun readSongFile(uri: Uri): String {
-        try {
-            return readSongFileUnsafe(uri)
-        }
-        catch(e: IOException) // Includes FileNotFoundException
-        { }
-        return "Cannot read file \uD83D\uDE1E" // sad emoji
-    }
-
-    private fun readSongFileUnsafe(uri: Uri): String {
-        val stringBuilder = StringBuilder()
-        contentResolver.openInputStream(uri)?.use { inputStream ->
-            BufferedReader(InputStreamReader(inputStream)).use { reader ->
-                var line: String? = reader.readLine()
-                while (line != null) {
-                    stringBuilder.append(line)
-                    line = reader.readLine()
-                }
-            }
-        }
-        return stringBuilder.toString()
-    }
+    private lateinit var song: Song
+    private val songViewModel: SongViewModel by viewModels()
 }
