+++ app/src/main/java/com/example/repertoire/SongRepository.kt
@@ -7,7 +7,11 @@ import android.database.Cursor
 import android.net.Uri
 import android.provider.OpenableColumns
 import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import java.io.BufferedReader
 import java.io.File
+import java.io.IOException
+import java.io.InputStreamReader
 
 class SongRepository(
     private val resolver: ContentResolver,
@@ -30,6 +34,10 @@ class SongRepository(
         return songDao.getAllLive()
     }
 
+    fun getSongContentLive(): LiveData<String> {
+        return songContent
+    }
+
     suspend fun remove(uri: Uri) {
         resolver.releasePersistableUriPermission(uri, Intent.FLAG_GRANT_READ_URI_PERMISSION)
         songDao.delete(uri.toString())
@@ -50,5 +58,35 @@ class SongRepository(
         return name
     }
 
+    fun setSongContent(song: Song) {
+        val chords = readSongFile(Uri.parse(song.uri))
+        songContent.value = "${song.name} $chords"
+    }
+
+
+    private fun readSongFile(uri: Uri): String {
+        try {
+            return readSongFileUnsafe(uri)
+        }
+        catch(e: IOException) // Includes FileNotFoundException
+        { }
+        return "Cannot read file \uD83D\uDE1E" // sad emoji
+    }
+
+    private fun readSongFileUnsafe(uri: Uri): String {
+        val stringBuilder = StringBuilder()
+        resolver.openInputStream(uri)?.use { inputStream ->
+            BufferedReader(InputStreamReader(inputStream)).use { reader ->
+                var line: String? = reader.readLine()
+                while (line != null) {
+                    stringBuilder.append(line)
+                    line = reader.readLine()
+                }
+            }
+        }
+        return stringBuilder.toString()
+    }
+
     private val songDao = db.songDao()
+    private val songContent = MutableLiveData<String>()
 }
